requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
python-multipart==0.0.17
Jinja2==3.1.4
httpx==0.27.2
lumaai>=0.1.0
python-dotenv==1.0.1
